refactor(test): split testing into per-model/layers files
