+++ machine_learning/src/test/mod.rs
@@ -1,0 +1,18 @@
+mod test_sequential_conv_dense;
+mod test_sequential_dense;
+
+use rand::Rng;
+
+fn gen_params_grads(server_sizes: &[usize]) -> Vec<(Vec<f32>, Vec<f32>, Vec<f32>)> {
+    let mut rng = rand::rng();
+    server_sizes
+        .iter()
+        .map(|&size| {
+            (
+                (0..size).map(|_| rng.random_range(-0.5..0.5)).collect(),
+                vec![0.; size],
+                vec![0.; size],
+            )
+        })
+        .collect()
+}
